+++ src/main/asciidoc/override/json.adoc
@@ -119,7 +119,7 @@ You use {@link io.vertx.core.json.JsonArray#encode} to encode the array to a Str
 
 === JsonCodec
 
-If you want to define a custom way to encode and decode Json, you can define a {@link io.vertx.core.json.JsonCodec} for it.
+If you want to define a custom way to encode and decode Json, you can define a {@link io.vertx.core.spi.json.JsonCodec} for it.
 A `JsonCodec` for type `X` is a concrete class that implements the interface `JsonCodec<X, J>`, where `J` must be any of these:
 
 * `JsonArray` or `JsonObject`
@@ -128,4 +128,4 @@ A `JsonCodec` for type `X` is a concrete class that implements the interface `Js
 * `Boolean`
 
 These `JsonCodec`s can be used in several components of Vert.x stack, like in vertx-web-client, using the SPI
-`io.vertx.core.json.JsonCodec`.
+`io.vertx.core.spi.json.JsonCodec`.
