+++ src/main/java/io/vertx/core/spi/json/JsonCodec.java
@@ -1,4 +1,7 @@
-package io.vertx.core.json;
+package io.vertx.core.spi.json;
+
+import io.vertx.core.json.JsonArray;
+import io.vertx.core.json.JsonObject;
 
 /**
  * This interface represents the Vert.x json representation of TARGET_TYPE. <br/>
