+++ src/test/resources/META-INF/services/io.vertx.core.spi.json.JsonCodec
